refactor: modularize main script
